ai-service/requirements.txt
langchain-nebius
langchain
langgraph
langchain-tavily
google-search-results
python-dotenv
pydantic
rich
